[Refactoring ⚙️] PostListCard Component 분리
